build(webpack): move CopyUniconsPlugin into own file
